+++ app/src/main/kotlin/org/fossify/notes/dialogs/NewChecklistItemDialog.kt
@@ -46,6 +46,7 @@ class NewChecklistItemDialog(
             .setNegativeButton(org.fossify.commons.R.string.cancel, null)
             .apply {
                 activity.setupDialogStuff(view, this, R.string.add_new_checklist_items) { alertDialog ->
+                    alertDialog.setCanceledOnTouchOutside(false)
                     alertDialog.showKeyboard(titles.first())
                     alertDialog.getButton(BUTTON_POSITIVE).setOnClickListener {
                         activity.config.addNewChecklistItemsTop = binding.settingsAddChecklistTop.isChecked